refactor : GameStart 패키징
