+++ app/src/main/java/com/ursssu/unscramble/presentation/gamestart/GameStartViewModel.kt
@@ -1,4 +1,4 @@
-package com.ursssu.unscramble.presentation
+package com.ursssu.unscramble.presentation.gamestart
 
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
